requirements.txt
requests>=2.28.1
sqlalchemy>=1.4.40
psycopg2-binary>=2.9.3
python-dotenv>=0.20.0
tqdm>=4.64.0
click>=8.1.3
pandas==2.3.1
